2/6/2024

+++ app/src/main/java/com/kan/dev/familyhealth/ui/fragment/StatisticFragment.kt
@@ -321,7 +321,7 @@ class StatisticFragment : BaseFragment<FragmentStatisticBinding>(),IRecentListen
                     Log.d("Kan", "Month :$months")
                     if (months == currentMonth) {
                         x = time.substring(time.lastIndexOf('.') + 1).toFloat()
-                        y = it.bmi.toFloat()
+                        y = it.bmi
                         values.add(Entry(x!!, y!!))
 
                         groupedByX.clear()
